requirements.txt
beautifulsoup4==4.12.3
pandas==2.2.1
python-dotenv==1.0.1
Requests==2.31.0
selenium==4.18.1
spotipy==2.23.0
seaborn
scikit-learn
scipy
matplotlib